Move RemoteOptions struct to remote_provider_traits
