+++ src/rust/engine/fs/store/src/remote/reapi.rs
@@ -29,9 +29,7 @@ use tokio::sync::Mutex;
 use tonic::{Code, Request, Status};
 use workunit_store::{Metric, ObservationMetric};
 
-use crate::RemoteOptions;
-
-use remote_provider_traits::{ByteStoreProvider, LoadDestination};
+use remote_provider_traits::{ByteStoreProvider, LoadDestination, RemoteOptions};
 
 pub struct Provider {
   instance_name: Option<String>,
