+++ src/rust/engine/process_execution/remote/src/remote_cache/base_opendal_tests.rs
@@ -12,7 +12,7 @@ use opendal::services::Memory;
 use process_execution::Context;
 use prost::Message;
 use protos::gen::build::bazel::remote::execution::v2 as remexec;
-use store::remote::RemoteOptions;
+use remote_provider_traits::RemoteOptions;
 
 use super::base_opendal::Provider;
 use super::ActionCacheProvider;
